Add support for unicode
A new action is added to input unicode. In the cfg file, unicode inputs
must be explicitly labelled as a `unicode` action, as opposed to
automatically being able to be parsed as a regular key like f2 or pgup.
This is because unicode outputs are special.

In Windows the keyboard hook supports inputting unicode characters
directly. Not all applications work properly — one notable example I
found from my own use cases is that the fancy new Windows terminal of
all things doesn't handle this correctly.

In Linux, unicode characters are input by typing regular characters in
the sequence: `C-S-u <hex code, e.g. 2f83> ret`. This is also not
guaranteed to work in all applications, but I've tested that it works in
the default xubuntu 20.04 terminal and Firefox.