Attempt fix for alt/win held on kanata startup
This commit attempts to fix #5 by sending a key release for all mapped
OsCode's on startup. This is done because kanata is not aware of any
keys that may be known as "held" before startup, so just for safe
measure, release everything. Upon initial testing this seems to work,
but #5 should be kept open for a while longer to fully confirm that the
fix works.