Add mouse support
This commit adds mouse support for left,right, and middle mouse clicks.
Support for mouse fwd/back buttons is not added because I couldn't get
them to work either in Linux or Windows. Instead the fwd/back keys are
used on the example mouse layer, which seems to behave exactly the same.

Scrolling is also not added because it's more complicated to implement
and I personally don't need it. I did try for some time but couldn't get
it to work in Linux.