Print layer as info log when layer changes
The motivation behind this change is for getting accustomed to new
layout changes. The log may be of benefit so that the user can see what
layout they're currently on and what keys are associated with it.

The info-level logs are currently empty apart from initialization, so
may as well make it useful for something.

One current limitation is that any comments in the original layout file
won't exist in the printed text.

Additional changes:

- increment version
- refactor handle_time_ticks into multiple fns
- change to new kanata-keyberon version for pub fn current_layer