Add handling for quoted string
Support for a quoted string as an SExpr atom is added in order to
support quoted strings as CLI arguments.

Some current weaknesses:
- closing parens are not ignored while inside a quoted string
- quoted strings replace all instances of one or more whitespace
  characters and newlines with a single space

Fixing the weaknesses above is a non-goal of kanata at this time. For
complicated commands, the user should either write a script/program and
call that from kanata rather than adding a complicated command right
within kanata.