Add parsing for cmd action
This commit adds parsing for the cmd action. In addition, enabling
parsing for the cmd action is gated by a defcfg entry so that the user
must opt into the potentially dangerous functionality.

A good chunk of the parsing code is refactored in order to more easily
pass around the parsed state including whether or not the cmd action has
been enabled.