Implement cmd execution
This commit implements the execution of the cmd action. The cmd is
executed in another thread so that the keyboard processing thread does
not get blocked.

The cmd action is gated behind conditional compilation as well, so that
a user of kanata can choose not to have the capability be possible at
all, even if the configuration might be present.