Make multi action work with non-keyberon actions
Due to how the keyberon library works, only one Custom action will be
returned per tick. This means that for a multi action, kanata will only
ever get back one Custom action from the keyberon library. As a first
cut attempt to work around the issue, this commit adds Multi versions of
the cmd, unicode, and mouse actions which are handled differently.

An alternative solution could be to make the custom action used with
keyberon be a `&'static [&'static CustomAction]`. However this would
have required more changes and I felt that the functionality that this
enables wouldn't be of any use in practice.