Use evdev to create and write device
The evdev crate is added and these dependencies are dropped:
- libc
- uinput-sys
- evdev-rs

Replace uinput-sys with evdev because evdev is a wrapper library for
evdev devices that provides interfaces to create uinput devices and send
events. Using evdev is less error-prone than accessing uinput directly,
and should be considered for new software[1]. evdev is also needed to
create a symlink in /dev/input/by-id/

libc is dropped since its struct definitions are no longer required.

evdev-rs is replaced by evdev so that the system library libevdev is no
longer a dependency.

[1]: https://www.kernel.org/doc/html/latest/input/uinput.html