Revert changes to evdev-rs and uinput crates
This reverts commits:
- bcbb5ec895e274b22ae060a1299e585560af2f95.
- 2e66b323efaadf352ca7471165490d65028247e4.

The reason for doing this is that evdev-rs=0.5.0, evdev-sys=0.2.4 have
a bug where they aren't cached properly and get recompiled on every
build/test/check. This is degrades the development workflow on Linux.

The issue has been fixed but there is no version published to crates.io
that includes the fix.

See: https://github.com/ndesh26/evdev-rs/issues/88