feat(tcp): add keep-alive, split client/server messages
This commit sets a 30-second keep-alive on opened TcpStreams (using
net2::TcpStreamExt as keep-alive functionality is not currently in the
standard library) and splits the previous EventNotification enum into a
ServerMessage and a ClientMessage enum respectively.

I have changed the semantics between ServerMessage and ClientMessage
slightly so that ClientMessage variants are imperative requests that
align closer with the names of the handler functions, ie. ChangeLayer
calls fn change_layer().

Whenever a client's TcpStream cannot be written to, either because it
has notified the server of a disconnect or because it has failed the
keep-alive, it will be removed from the connections HashMap on the
TcpServer struct.

re #47