fix(poll): allow poll to be interrupted
From manual testing it was discovered that the call to mio's poll
function sometimes returns with OS code 4: syscall interrupted. This was
seen when testing the multi action with multiple mouse buttons and
multiple cmds.

This commit fixes the issue by logging and otherwise suppressing poll
errors in the event loop.