refactor: send normal keys before handling custom actions
This change of ordering is done because a user reported that the multi
action along with a combination of keys and mouse buttons did not have
the expected order for their use case.

I can't currently imagine a scenario where a mouse click should occur
immediately before a button press so I will guess that it's fine to make
this change.