feat: add mouse tap action
This feature is added because on Windows, the behaviour of Ctrl+LClick
to click on a link in a browser depends on the release timing rather
than the click timing. If the LClick is released before Ctrl, a new tab
opens but if Ctrl is released first, the link is followed without a new
tab (unless it always opens a new tab).

However, a user reported a bug that affected their workflow they needed
Ctrl to occur before MClick (see previous commit). The naive moving of
custom action handling to occur after regular key handling breaks the
Windows Ctrl+LClick example workflow, because Ctrl gets released first.

The solution to this is the "tap" mouse actions, which if used will mean
that using (multi lctl mltp) will occur in the action sequence of:

    press key->press lctl->click left->unclick left->release key->release lctl

This sequence results in the correct behaviour for Windows Ctrl+Click to
open a new tab and hopefully the correct behaviour for the user.