feat: add alias for layer-toggle: layer-while-held
This commit adds an alternate action name for layer-toggle:

    layer-while-held

This is added because the word "toggle" does not quite precisely
describe what the action actually does.

A miscellaneous change included in this commit is adding some missing
action names in an error message and changing the error message's format
to group related commands together.