Static lifetime for the low-level keyboard hook
Fixes a soundness issue I found in the original
[kbremap](https://github.com/timokroeger/kbremap/blob/main/src/keyboard_hook.rs)
code.

Drop code is not guaratueed to run (e.g. when using `mem::forget`).
Thus we cannot rely on drop unregistering the hook.
Make the hook abstraction require a static lifetime for the closure to
ensure crash free operation in any case.

Program to reproduce the issue:
``` Rust
fn main() {
    {
        let mut pressed_keys = Vec::with_capacity(1);
        let hook = KeyboardHook::set_input_cb(|ev| {
            pressed_keys.push(ev);
            false
        });
        mem::forget(hook);
    }

    unsafe {
        let mut msg = mem::zeroed();
        GetMessageW(&mut msg, ptr::null_mut(), 0, 0);
    }
}
```