Add 'pause/break' key and fix cmp and clr (#129)
The clear key exists in windows.rs, but was missing from linux.rs. Now the key works in Kanata under Linux, although I'm not actually sure if it works as i can't get xev to output anything when i press it.

The menu key being named 'compose' is confusing, so 'menu' as an alias has been added.
Another alias 'apps' is added to reflect VK_APPS in the Windows virtual-key codes.