fix: add defcfg item to process unmapped keys (#144)
This commit adds a new defcfg item and reverts the first attempted fix
for tap-hold-(press|release) not being triggered by keys that are not in
defsrc. The first fix was not correct because the order of keys was
incorrect.

Processing keys missing from defsrc is disabled by default. The reason
this is not enabled by default is because some keys may not work
correctly if they are intercepted. E.g. ralt/altgr on Windows using the
LLHOOK mechanism.