fix(windows): release all shift keys on non-physical shift release (#142)
Windows has some weirdness with arrow keys and the numpad where one is a
shifted version of the other. I think it's a historical issue since
physical arrow keys don't have this problem but injected ones do. Might
be a bug in SendInput or the LLHOOK mechanism. This is relevant because
it's likely related to the point below.

When injecting an arrow key while an injected lsft is held, Windows
sends a fake physical lsft release and then a fake physical lsft press
event. The kanata-injected arrow event is likely in between the two fake
physical events. I think Windows does this because at the OS level, it
thinks lsft needs to be released to be able to send an arrow key instead
of a numpad key. So since the injected lsft is active in the Windows OS,
Windows sends the fake lsft events without the physical keyboard being
involved.

The problem is that the key Windows sends is the physical lsft key. This
results in kanata's state saying that the lsft key is pressed down
permanently until the physical lsft key is pressed and released. In the
kanata software, the state of the key mapped to lsft independent of the
state of the physical lsft key. Releasing the key remapped lsft, will
not release the state of the lsft key.

The fix for this issue is that on Windows, non-physical lsft/rsft
releases will release all active lsft/rsft states. On Linux, this code
does not run at all. My assumption is that the use case of pressing two
of the same shift at the same time is not depended on by anyone and is
not useful, so this shouldn't negatively affect anyone.