feat: add cancellable macro
The use case motivating this feature is to be able to hold a key for
different amounts of time and outputting different characters with
visual feedback on what's getting outputted. It is similar to tap-dance
but instead of tapping a key for varying number of repeats, the key is
held for a varying amount of time to get the different outputs.