feat: allow some special actions in macro (#182)
This commit enhances the `macro` and `macro-release-cancel` actions by
allowing some other special actions to be used inside. Some actions,
e.g. `tap-hold` don't make much sense to put in a macro, so are not
supported.

The precise list of actions allowed within `macro` are those that
correspond to an action in the `CustomAction` enum.

At the time of writing, the list of useful actions supported in macros
are:

- cmd
- unicode
- mouse actions

There are more actions that are technically allowed, but seem like they
don't make much sense to put inside of a macro. Thus they are not
documented:

- fake key actions
- sequence leader
- live reload
- repeat