fix: add multi key and multi action to key outputs (#190)
This commit fixes a bug where `multi` and chorded keys like `S-b` were
not generating key outputs for the purposes of repeat. The new code
fixes the issue with hopes to prevent further regressions by changing
the keyberon `Action` to be exhaustively matchable and explicitly
handling all cases instead of using a catch-all `_ => { ... }` branch.

Another underlying bug was uncovered where key repeat for chords like
`S-b` were inconsistent due to iterating over a HashSet. The bug was
that sometimes the Shift would be repeated, and other times the B
repeated due to the random iteration order over a HashSet. The code now
uses a `Vec` inside of the `KeyOutputs` and also adds layer info to it
to improve key repeat consistency.

This opportunity is also taken to switch to `rustc-hash` which uses
FxHash to reduce hashing cost compared to the default SipHash — which is
cryptographically secure but more expensive to compute.