feat: be more lenient for missing linux device
This commit changes the linux device opening code to only generate
warning logs when a device cannot be opened. An error is only returned
when no devices can be successfully opened.

A minor improvement to the main function is included. The main function
now exits when an error is encountered and the error logging clarity is
improved by quoting the error.