feat(win-interception): implement mouse btns in defsrc
This commit implements mouse buttons in defsrc for Windows with the use
of the interception driver. It includes a new defcfg item which states
which mouse hardware ID to intercept. Since the intended use case is for
laptops that have builtin mouse buttons like Lenovo Thinkpads, only a
single hardware ID is allowed for now.

This commit also fixes a bug in the parsing which probably only affected
the `cmd` action. The bug is that quoted strings keep their quotes when
parsed as an atom by the S-expression parser. This was introduced in
v1.0.8. This behaviour is not changed; instead, the quotes are trimmed
when parsing `cmd` and `defcfg`. If one day the parser is properly fixed
to remove the leading and trailing quotes from atoms, then the trimming
code should still be fine; `trim_matches` will be fine with or without
the quotes.