feat!: add custom key labels in cfg
This commit adds a new configuration item `defcustomkeys`. This allows a
user to define custom keys that can be used without aliases in `defsrc`,
`deflayer` and other config items.

The documentation is updated to hopefully encourage community-created
resources for non-US keyboard `defsrc` configurations.

A breaking change is done by removing `ì` from `str_to_oscode`. I
consider this fine since it is potentially buggy if it is there; other
locales could potentially have a different mapping for that key. It
should instead be defined in `defcustomkeys`. I also find it unlikely
that anyone is using it since its addition, and even then it may have
been mapped incorrrectly.

The key name `+` is changed to be a "default custom key". This is
because some locales have it as an unshifted key, but it was previously
mapped to the numpad plus key. It can be overridden if desired, but
otherwise will behave the same as before to maintain backwards
compatibility. This is treated differently from `ì` due to the length of
time it's been possible to use as a mapping as well as my percieved
likeliness of it being used.