feat(linux): improve behaviour of device discovery (#219)
This commit adds use of inotify to respond properly to changes in
/dev/input to improve the responsiveness after a new device is created.

This commit also adapts code from keyd to read an input device's
keystate and only grabs when all keys are released.

An option is added that changes kanata's behaviour to not crash if no
devices are able to be registered. Since kanata is now able to
consistently deregister and reregister devices, some may find this
behaviour desirable.