feat: send LayerChange from server when client connects
The purpose of this change is to allow a client to know what layer is
currently active when it connects to the server. When combining this
change with existing `LayerChange` notifications, a client will be able
to know what layer is active from the beginning to the end of the client
lifetime.

The `as_bytes` code for ServerMessage is simplified because I don't see
any normal reason why serialization of this type should fail. A test is
also added to check this. An example instance of why a different type
might fail is if the key of a HashMap does not serialize into a "string"
type.