feat(sequence): add input mode configuration
This commit adds a configuration to customize the input mode for
sequences via a defcfg item: `sequence-input-mode`. One can choose
between:

- visible-backspaced
- hidden-suppressed
- hidden-delay-type

The behaviours of these three options are documented in kanata.kbd and
config.adoc appropriately. The default option is hidden-suppressed,
which matches the old, non-configurable behaviour for sequence input, so
as to not break someone who may be depending on or is used to this
behaviour.