fix(tap-hold-eager): change assert and interruption
This commit changes the keyberon action assertion. The intent of the
old assertion was as a bug check in the code and configuration. If a
waiting state exists, e.g. tap-dance or tap-hold, then the code should
not be performing any actions until that state is done processing.
However, the `Custom` action type is special because it has minimal
impact on the rest of the keyberon state, and so is fine to allow in.
This enables a previously crashing use case, which is fine to not crash
on.

In addition, fake keys will no longer interrupt the tap-dance-eager
state. This is so fake key actions like on-release-fakekey can be used
while in the middle of a tap-dance-eager sequence, which is a requested
use case.