refactor: free allocated memory on layout drop
This commit fixes the months-old known issue where live reload leaks
memory. This is done by tracking "leaked" allocations and freeing them
when the layout is dropped.

In the case that a configuration error is found and the layout is not
created, the code will clean up those unused allocations as well.

Naturally, this makes use of unsafe. The code has been tested in
valgrind with no reported memory violations or leaks.