refactor: free allocated memory on layout drop (#272)
This commit fixes the months-old known issue where live reload leaks
memory. This is done by tracking "leaked" allocations and freeing them
when the layout is dropped.

In the case that a configuration error is found and the layout is not
created, the code will clean up those unused allocations as well. The
keyberon code is changed to use explicit lifetimes and the Cmd action is
changed to use a Vec to catch potential UB.

This makes use of unsafe. The code has been tested in valgrind with no
reported memory violations or leaks. The unsafe should be isolated to
the `cfg` module.

The `cfg` module makes use of `'static` lifetimes to simplify the code,
but restricts users of the generated layout to a shortened lifetime. The
Cmd variant of CustomCommand is also changed to use a Vec for safety
purposes, since before it was using a static lifetime.