fix: disallow multiple waiting actions (#285)
This commit fixes the configuration parsing to reject multiple waiting
actions to be combined in a single `multi`. (tap-dance, tap-hold-*,
chord). This is done because such an action will always crash at
runtime. The keyberon code cannot handle this type of action and such
has an assertion to crash if encountered.