feat: add configurable replay delay
This commit allows an option to configure dynamic macro replay delay.
The reason this exists is that the default 5 ticks of delay may not be
sufficient to the keyberon state machine to finish a state transition
between actions, in which case the user might need to increase the delay
value.