dep: cleanup, panic = "abort", update clap
This commit cleans up dependencies and dependency features with the goal
of reducing compile times and code size. In the same goal, the line:

    panic = "abort"

is added to Cargo.toml. Currently kanata does not rely on unwinding
anywhere, so this seems fine to do.

The clap crate is updated to the latest, for no particular reason other
than keeping it up to date.