feat(linux): add mouse move support #311
Add support for linux mouse movement events with the evdev crate.

This commit includes a minor performance refactor. From reading
documentation, using `new` when sending to the kernel is fine because
the kernel will fill in the default value of 0 for the time with a
proper value for any reading devices. This refactor is done to remove
whatever call is used to fill in the time value, for very minor,
probably invisible, performance gain.