fix(wintercept): detect empty ch, use unbounded ch
The std channels are missing the is_empty check. This was needed to
avoid having a leftover event in the channel before blocking. This
commit adds a workaround to allow checking whether the channel is empty.

The interception channel is changed to unbounded in order to handle
dynamic macro replay, which can overwhelm a bounded channel.