fix: release lock before sleeping
Otherwise the lock is held almost continously while any kind of keyberon timer
is active. That (at least on Linux) prevents events which bypass kanata (such
as mouse movement) from being forwarded in a timly manner (resulting in delayed
and low-frequency mouse movement) because the event loop thread will be unable
to acquire the lock it needs to output events.