refactor(wintercept): use thread_local IC in oskbd
This commit refactors the interception code to use a second instance of
an Interception client in `oskbd`. Initially I had used channels to
communicate from `oskbd` to `kanata` because `Interception` is not Send,
so could not be stored within `KbdOut`. However, it did not occur to me
at the time to use thread local storage to persist an Interception
client within the `oskbd` module. Now that I realize that is possible,
it simplifies the interception variant of kanata greatly and also
removes the terrible can-block detection - the interception event loop
can simply always do a blocking call.