refactor: move some handle_repeat code and add comment
When re-reading this I got confused about why cur_keys was being used
and why it was being cleared. I hid it within handle_repeat to mitigate
it from confusing me again. I also added a clarifying comment to explain
what it's being used for and why it's ok to clear.