refactor: reduce size of KanataAction (#321)
This commit adds an extra layer of indirection on all slices — which
means they are regular pointers instead of fat pointers — to be able to
reduce the size of KanataAction to 2 pointers instead of 3 pointers.

This has a performance penalty of an additional indirection for
dereferencing slice actions. It also makes the code for creating actions
even uglier than it already is. However, it has the performance gain of
improving cache locality and reducing total memory footprint by 1/3.
Overall it is not clear whether this is a win.