refactor: remove dead code, review unwraps
The code not useful to kanata in the keyberon crate is removed.

Some unwraps are removed while the rest that are not in tests are
changed to expect for clearer error messages in case of an unexpected
problem.

The cfg tests are moved to a new file for easier future unwrap review.