fix: revert refactor(wintercept): use thread_local IC in oskbd (#341)
This reverts commit 6af336073d393bcbb4d6847939d54308c0c4329c.

Lately on one of my machines I have encountered issues with the
interception driver mechanism dying on me even though kanata appears to
be operating just fine. Kanata is able to read the keyboard events just
fine and logs writes of the expected keys to the output. However, the
key output doesn't work and nothing happens in the OS. I would guess
it's caused by use of an interception handle in two different threads.