feat: tap-hold-release-keys (#343)
This commit adds another tap-hold variant that makes use of keyberon's
`HoldTapConfig::Custom` variant for the tap hold configuration. The
custom variant is used to activate an early tap when any key in the
given list is pressed. The goal of this variant is to reduce rolling
errors on the same hand when using home-row tap-hold keys.

An unrelated fly-by fix to the minimal.kbd configuration file is added.
There was some unintended leftover text that is now deleted.