doc: adjust build/install instructions and release template
Add section on feature flags and shortens the README by default using
collapsed HTML in the DIY section. A note on making changes to the
keyberon code is also added to the contributing section. The text is
reorganized to put the running-in-background text outside of the
collapsed sections.

This commit also tries out the use of semantic newlines in the README as
opposed to sticking with an 80 character width.

The release template is changed to use collapsed HTML as well.

Some minor adjustments to the config samples are done.

The "Features" section is renamed to "Notable features" since it's not a
complete list.