fix: reduce blast radius of windows sft-arrowkey workaround (#414)
This commit makes changes to the blast radius of the workaround for
issue #138:

> Shift and arrows interact weirdly on a layer

Discovered in this issue was a problem where activating shift on keys
other than the physical shift keys, and then also pressing arrow keys,
caused stuck-shift behaviour. The fault is, as is typical, Windows'
weird behaviour. The workaround added was to release all shift key
states if a single shift key state was found to no longer exist, when
comparing the previous keyberon state to the current one.

In some cases, this release of shift causes an earlier release of shift
than is desired, because some other key activated shift. This commit
reduces the blast radius of the shift state releases to only states with
the physical lsft key. In hindsight, Windows' weird behaviour only
would have affected the physical lsft key's coordinate anyway, so this
is a good change to make. In addition, this issue only affects the
LLHOOK mechanism and not Interception, so the code is conditionally
compiled out for the Interception driver as well.