feat: improve error message when list action is not in a list (#565)
Before this commit, the error message when a list action was not
parenthesized was:

    Unknown key/action/variable: {ac:?}

This is not helpful and confusing because the action is valid and
exists, but just needs to be in a list. This commit improves the error
message to state that the action must be in parentheses.