refactor: use bounded channel (#579)
This commit attempts to fix an issue with excessive CPU and memory
usage. At the time there are no steps to reproduce, but one suspicion is
masses of rogue input events from the kernel. Thus the code is changed
to use bounded channels with a reasonably large size (100) that I
wouldn't expect to get filled up.