feat!: conditionally compile tcp server code (#680)
The motivation for this commit is to enable savings on binary size if
desired.

This commit turns the TCP server code into an optional feature. The
feature is enabled by default for backwards compatibility. Turning off
default features saves 81920 bytes from the stripped binary when
compiling with release profile when testing on my system.