feat!(dynamic-macro): record and simulate delays by default (#685)
This is a breaking change in behaviour of dynamic macros. I'm not sure
that the anyone would really want the old behaviour though. It seems odd
that someone might be dependent on it.

Kanata defaults to the new behaviour but can be reverted if desired. I
may choose to remove the `defcfg` option though, and **only** have the
new behaviour. For now the option is left undocumented, so that I can
argue that it is not a breaking change if I choose to remove the option
in the future, since the option was never documented ;)