fix(win-llhook): make read-scancode actually work (#871)
The previous commit did not work as intended because converting scancode
to vk used the layout active at the time of kanata startup - which is
the behaviour of `MapVirtualKeyA`. But this is not what we want. For
internal use we want a constant `scancode<->vk` mapping regardless of
layout, to allow `scancode->vk->scancode` to work correctly, where the
intermediate vk is what's used for kanata's processing, in between the
hook input and SendInput output.