refactor(tcp): use stream reader, wake up on all msg types
Using a stream reader prevents incomplete reads from unnecessarily
interrupting a connection, though in practice incomplete reads are
highly unlikely to happen because client messages are typically short.

The kanata state machine is also now woken up on all client messages.
The reason is that I was noticing a delay in layer changing, where it
would only happen after a key that was being processed by kanata was
typed. It should not be too costly to do this, even for client messages
that don't need to wake up the processing loop, because the processing
loop should quickly go back to idle if it wasn't necessary. It is
simpler and clearer in the code to always trigger the wakeup instead of
special-casing it for only events that truly need it.