feat(win): embed a manifest to allow using kanata in elevated windows (#946)
Add a manifest to the kanata.exe Windows binary so that external tools
can overwrite it to make it work in elevated contexts. This manifest
doesn't have `uia` enabled on its own as this doesn't work with an
unsigned binary

So I've also added AutoHotkey scripts that sign the binary (and enable
`uia`).