Stop visible-backspace from pressing normalized mods
When `visible-backspace` is enabled and a sequence containing a shifted
key is activated with RSFT, the captured `Press` would be normalized to
LSFT for comparison against the sequence trie before being re-used by
`visible-backspace`. Since the uncaptured `Release` event is still for
RSFT, we get a mis-match.

I've just scoped that normalization down to the variable
`pushed_into_seq`, which already existed for scoping the encoding of
modifiers into keycodes. The local variable `osc` in the following
`invalid termination` section gave me pause but is an unrelated
shadowing of the term, so there are no other uses that could conflict
with this change.