fix(repeat): check event code itself for repeat output
This commit fixes key repeat in Linux TTY and Windows when a key on the
base layer is transparent or unmapped, but process-unmapped-keys is
enabled.

This functionality was broken in commit 089853a which changed how layers
worked by removing the duplicated base-version and while-held-version of
layers, for the benefit of reduced memory consumption.