feat(win): add kanata native Windows GUI tray app (#990)
Add a `gui` feature allowing compiling kanata as a native tray-only
Windows GUI application. This greatly simplifies launching the app on
startup instead of resorting to hacks like this:

	https://github.com/jtroo/kanata/discussions/193

It gets rid of the console window and also allows UIAccess to work in
elevated apps. There is a kanata-tray project which helps with the
former, but doesn't allow UIAccess.

	issue: (https://github.com/rszyma/kanata-tray/issues/19)

It also supports custom icon indicator per config and per layer within
the same config, as well as (re)loading configs. Other configs are
passed as cli args in the `.lnk` file

It also supports being launched from a console as a regular cli app, so
it's possible to replace the main `kanata.exe` with this or let it be
some separate `kanata-tray.exe` file

It also opens an avenue to more interesting ideas, but will add those in
a separate issue https://github.com/jtroo/kanata/issues/991