refactor(win): adjust default feature flags
This commit adds win_sendinput_send_scancodes as a default feature. The
feature has had some time to be tested and should be good enough to
replace vkey output by default. With this change, anyone compiling the
default kanata binary on Windows without knowing about the difference
between vkey codes and scan codes and no awareness of winiov2 will
hopefully have a better experience than they used to.

The justfile line for compiling the legacy binary is adjusted as well so
that it compiles without the now-default feature and continues to use
vkey output, in case someone really needs it.