doc: fix nonexistent key names in fancy_symbols.md (#1172)
Update some outdated key names in the list of fancy symbols. For
example, the table says `page_down`, but this throws an error in kanata
because it actually expects `pgdn`.

It took me time to find the actual name (I had to use the symbol ⇟ while
I found it was `pgdn`), so I think it's better to put the actual key name
in the table. I presume that `page_down` was used in earlier versions
of kanata, which would explain the mismatch. I updated the ones I
noticed. Not all of them are fixed I think.