docs: note for PATH env in systemd service
as the PATH env in the example service limits to global paths of kanata, issues may happen when the binary is installed inside the user's home directory, as shown in issues #1153 and #1179. this commit attempts to remedy such problem by explaining how to include the path to locally installed kanata.