feat: allow for include files to be absent
Sometimes include files are not present (they might be host specific)
but the main configuration file should nevertheless reference them (as
this file is tracked in git). At the moment, included files need to be
present for kanata to start. This restriction is now loosened and
non-existing files are ignored and a info message is shown for such
cases.