Recommend high priority for Linux
This change recommends using a higher
(in fact, the highes) CPU and IO priority
to the kanata process.

While this is not such an obvious default
due to the mouse being by default captured by kanata,
the change still makes sense as keyboard input
is likely seen as a real-time priority for kanata users.

Additionally, I've removed the DISPLAY environment
variable from the service. It's not needed.
Kanata works entirely outside of Xorg.
Nor does it have a GUI that would need to be shown.