fix(tcp): do not skip layer names
The step_by call was originally done because there used to be two versions of the same layer; one which was the target of `layer-switch` and the other which was the target of `layer-while-held`. This has since been changed and there is only one version of each layer now, so the step_by needs to be removed now.