fix: run zippychord test only when feature enabled (#1687)
One can activate a specific set of features so that the
zippychord_sim_tests test is run without the zippychord feature being
enabled. This leads to a runtime error saying that the zippychord
feature was not enabled when trying to run the tests with this specific
feature set.

Add an additional cfg feature gate in front of the zippychord_sim_tests
module to prevent this failure.